+++ durihana/components/atoms/Header.tsx
@@ -1,11 +1,10 @@
 'use client';
 
+import { Button, Txt } from '@/components/atoms';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { signIn, signOut, useSession } from 'next-auth/react';
 import { useEffect, useState } from 'react';
-import Button from './Button';
-import Txt from './Txt';
 
 type Props = {
   title?: string;
